Data manager

+++ 690Project/Project/FileSaver.cs
@@ -3,10 +3,13 @@ namespace Project;
 using System.IO;
 public class FileSaver
 {
+    
     string fileName;
     public FileSaver(string fileName) {
         this.fileName = fileName;
+        if(!File.Exists(this.fileName)){
         File.Create(this.fileName).Close();
+        }
     }
 
     public void AppendLine(string line)
@@ -15,6 +18,6 @@ public class FileSaver
     }
     public void AppendData(TaskData data)
     {
-        File.AppendAllText(this.fileName, data.User + ":" + data.Category + ":" + data.Label + data.DueDate + data.Status + Environment.NewLine);
+        File.AppendAllText(this.fileName, data.User + "-" + data.Category + "-" + data.Label + "-" + data.DueDate + "-" + data.Status + Environment.NewLine);
     }
 }
